sign compare

+++ source/client/player/input/Keyboard.cpp
@@ -28,7 +28,7 @@ void Keyboard::feed(KeyState state, int key)
 
 bool Keyboard::next()
 {
-	if (_index + 1 >= _inputs.size())
+	if ((size_t)_index + 1 >= _inputs.size())
 		return false;
 
 	_index++;
